Pull status from TBR server
Waiting for a commit to be pushed that will update the current status
of our services could potentially take a long time. If we host it on
our server, we can edit the file faster. Of course, if our server is
down, then this won't work, but if our server is down, we can't use
Gerrit to push a commit that updates the status of our services. With
the current status of our servers, it is best to just host this on our
own server.

Change-Id: I8103b4ce4877581f30ec97d347a16eaffb8f24d5